move errorscope to componentstates.kt
