+++ src/test/kotlin/org/jitsi/jibri/capture/ffmpeg/FfmpegCapturerTest.kt
@@ -34,6 +34,7 @@ import io.kotlintest.specs.ShouldSpec
 import org.jitsi.jibri.capture.UnsupportedOsException
 import org.jitsi.jibri.sink.Sink
 import org.jitsi.jibri.status.ComponentState
+import org.jitsi.jibri.status.ErrorScope
 import org.jitsi.jibri.util.JibriSubprocess
 import org.jitsi.jibri.util.OsDetector
 import org.jitsi.jibri.util.OsType
